requirements.txt
streamlit>=1.28.0
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python-headless>=4.6.0
matplotlib>=3.5.0
pandas>=1.3.0